+++ tests/__init__.py
@@ -1,1 +1,1 @@
-"""SDK Test Forge test suite."""
+﻿"""SDK Forge test suite."""